Move memory string formatting into utils
