+++ larq/utils.py
@@ -1,6 +1,27 @@
 from tensorflow.keras.utils import get_custom_objects
 
 
+def memory_as_readable_str(num_bits: int) -> str:
+    """Generate a human-readable string for the memory size.
+
+    1 KiB = 1024 B; we use the binary prefix (KiB) [1,2] instead of the decimal prefix
+    (KB) to avoid any confusion with multiplying by 1000 instead of 1024.
+
+    [1] https://en.wikipedia.org/wiki/Binary_prefix
+    [2] https://physics.nist.gov/cuu/Units/binary.html
+    """
+
+    suffixes = ["B", "KiB", "MiB", "GiB"]
+    num_bytes = num_bits / 8
+
+    for i, suffix in enumerate(suffixes):
+        rounded = num_bytes / (1024 ** i)
+        if rounded < 1024:
+            break
+
+    return f"{rounded:,.2f} {suffix}"
+
+
 def register_keras_custom_object(cls):
     """See https://github.com/tensorflow/addons/blob/master/tensorflow_addons/utils/keras_utils.py#L25"""
     get_custom_objects()[cls.__name__] = cls
